requirements.txt
ffmpeg-python
pytube
fireworks-ai
assemblyai
requests